stsWorkspace/py05/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
fastapi==0.115.14
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mysql==0.0.3
mysqlclient==2.2.7
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
